refactor(backoffice-v2): moves statistics sections to separate modules
